refactor: 검색 로직을 훅으로 분리하는 리팩토링
